Refactorizacion de analitica
