+++ app/src/main/java/com/jesuslcorominas/posts/app/data/analytics/events.kt
@@ -1,6 +1,6 @@
-package com.jesuslcorominas.posts.app.analytics
+package com.jesuslcorominas.posts.app.data.analytics
 
-import com.jesuslcorominas.posts.domain.analytics.AnalyticsEvent
+import com.jesuslcorominas.posts.data.source.AnalyticsEvent
 
 class ClickPostEvent(private val postId: Int) :
     AnalyticsEvent("ClickPost", mapOf("postId" to postId))
